nyay_ai/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv>=1.0.0
google-genai>=1.0.0

# Optional - install these if you want ML-powered retrieval (better answers):
numpy>=1.26.0
scikit-learn>=1.3.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
